+++ doc/changelog.rst
@@ -16,6 +16,7 @@ Added
 ^^^^^
 - The `Unknown resource type` request error keeps a reference to the faulty payload.
 - New `werkzeug` request engine for application development purpose.
+- New `AsyncSCIMClient` request engine. :issue:`1`
 
 Changed
 ^^^^^^^